requirements.txt
einops==0.6.0
mergedeep==1.3.4
numpy==1.21.6
pandas==1.4.2
PyYAML==6.0
scipy==1.8.1
tensorboard
torchmetrics
